Issue 263 (finish): Convert rest of servlets to Spring RestControllers.
