+++ app/src/main/java/org/apache/roller/weblogger/ui/rendering/util/WeblogFeedRequest.java
@@ -32,6 +32,7 @@ import org.apache.roller.weblogger.business.WeblogManager;
 import org.apache.roller.weblogger.config.WebloggerConfig;
 import org.apache.roller.weblogger.business.WebloggerFactory;
 import org.apache.roller.weblogger.pojos.WeblogCategory;
+import org.apache.roller.weblogger.ui.rendering.WeblogRequestMapper;
 import org.apache.roller.weblogger.util.Utilities;
 
 
@@ -46,8 +47,6 @@ import org.apache.roller.weblogger.util.Utilities;
 public class WeblogFeedRequest extends WeblogRequest {
     
     private static Log log = LogFactory.getLog(WeblogFeedRequest.class);
-    
-    private static final String FEED_SERVLET = "/roller-ui/rendering/feed";
     
     // lightweight attributes
     private String type = null;
@@ -83,7 +82,7 @@ public class WeblogFeedRequest extends WeblogRequest {
         log.debug("parsing path "+pathInfo);
         
         // was this request bound for the feed servlet?
-        if(servlet == null || !FEED_SERVLET.equals(servlet)) {
+        if(servlet == null || !WeblogRequestMapper.FEED_PROCESSOR.equals(servlet)) {
             throw new InvalidRequestException("not a weblog feed request, "+
                     request.getRequestURL());
         }
