+++ app/src/main/webapp/WEB-INF/jsps/tiles/footer.jsp
@@ -17,13 +17,12 @@
 --%>
 <%@ include file="/WEB-INF/jsps/taglibs-struts2.jsp" %>
 
-
 <% request.setAttribute("version",
       org.apache.roller.weblogger.config.WebloggerConfig.getProperty("weblogger.version", "Unknown")); %>
 <% request.setAttribute("revision",
       org.apache.roller.weblogger.config.WebloggerConfig.getProperty("weblogger.revision", "Unknown")); %>
 
-<img src='<c:url value="/images/tinyfeather.png"/>' alt="ASF logo" style="vertical-align:middle" />
+<img src='<s:url value="/images/tinyfeather.png"/>' alt="ASF logo" style="vertical-align:middle" />
 <s:text name="footer.productName" >
     <s:param value="#request.version" />
     <s:param value="#request.revision" />
