+++ app/src/main/webapp/WEB-INF/web.xml
@@ -159,36 +159,6 @@
     </listener>
 
     <servlet>
-        <servlet-name>FeedServlet</servlet-name>
-        <servlet-class>org.apache.roller.weblogger.ui.rendering.servlets.FeedServlet</servlet-class>
-        <load-on-startup>5</load-on-startup>
-    </servlet>
-
-    <servlet>
-        <servlet-name>MediaResourceServlet</servlet-name>
-        <servlet-class>org.apache.roller.weblogger.ui.rendering.servlets.MediaResourceServlet</servlet-class>
-        <load-on-startup>5</load-on-startup>
-    </servlet>
-
-    <servlet>
-        <servlet-name>SearchServlet</servlet-name>
-        <servlet-class>org.apache.roller.weblogger.ui.rendering.servlets.SearchServlet</servlet-class>
-        <load-on-startup>5</load-on-startup>
-    </servlet>
-
-    <servlet>
-        <servlet-name>PlanetFeedServlet</servlet-name>
-        <servlet-class>org.apache.roller.weblogger.ui.rendering.servlets.PlanetFeedServlet</servlet-class>
-        <load-on-startup>7</load-on-startup>
-    </servlet>
-
-    <servlet>
-        <servlet-name>PreviewServlet</servlet-name>
-        <servlet-class>org.apache.roller.weblogger.ui.rendering.servlets.PreviewServlet</servlet-class>
-        <load-on-startup>9</load-on-startup>
-    </servlet>
-
-    <servlet>
         <servlet-name>SpringDispatcherServlet</servlet-name>
         <servlet-class>org.springframework.web.servlet.DispatcherServlet</servlet-class>
         <init-param>
@@ -198,38 +168,17 @@
     </servlet>
 
     <servlet-mapping>
-        <servlet-name>FeedServlet</servlet-name>
-        <url-pattern>/roller-ui/rendering/feed/*</url-pattern>
-    </servlet-mapping>
-
-    <servlet-mapping>
-        <servlet-name>MediaResourceServlet</servlet-name>
-        <url-pattern>/roller-ui/rendering/media-resources/*</url-pattern>
-    </servlet-mapping>
-    
-    <servlet-mapping>
-        <servlet-name>SearchServlet</servlet-name>
-        <url-pattern>/roller-ui/rendering/search/*</url-pattern>
-    </servlet-mapping>
-
-    <servlet-mapping>
-        <servlet-name>PlanetFeedServlet</servlet-name>
-        <url-pattern>/planetrss/*</url-pattern>
-    </servlet-mapping>
-
-    <!-- Preview Servlets -->
-    <servlet-mapping>
-        <servlet-name>PreviewServlet</servlet-name>
-        <url-pattern>/roller-ui/authoring/preview/*</url-pattern>
-    </servlet-mapping>
-
-    <servlet-mapping>
         <servlet-name>SpringDispatcherServlet</servlet-name>
         <url-pattern>/roller-ui/authoring/rest/*</url-pattern>
+        <url-pattern>/roller-ui/authoring/preview/*</url-pattern>
         <url-pattern>/CommentAuthenticatorServlet</url-pattern>
         <url-pattern>/roller-ui/rendering/page/*</url-pattern>
         <url-pattern>/roller-ui/rendering/comment/*</url-pattern>
         <url-pattern>/roller-ui/rendering/trackback/*</url-pattern>
+        <url-pattern>/roller-ui/rendering/feed/*</url-pattern>
+        <url-pattern>/roller-ui/rendering/media-resources/*</url-pattern>
+        <url-pattern>/roller-ui/rendering/search/*</url-pattern>
+        <url-pattern>/planetrss/*</url-pattern>
     </servlet-mapping>
 
     <session-config>
